fix: keep location follow from fighting demos or losing the puck
Remember location state outside style composition, apply heading on the first fix, and make ferry follow and location follow exclusive.
